requirements.txt
discord.py
aiohttp
asyncpg
Pillow
python-dotenv
